sdk-python/pyproject.toml
[project]
name = "vouch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pooja Garg", email = "pooja.brown@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "pyyaml>=6.0",
]

[project.scripts]
vouch = "vouch:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
